requirements.txt
scikit-learn
numpy
pandas
pypdf
langchain_community
langchain
chromadb
rank_bm25
bitsandbytes
accelerate
torch
torchvision
transformers
sentence-transformers
weaviate-client
langchain_groq
langchain-groq
langchain-community
langgraph
langchain-openai
langchain_huggingface
langchain-core
langchain_chroma
langchain-classic
python-dotenv
pydantic
joblib
streamlit
tiktoken
faiss-cpu
Unstructured
unstructured[pdf]